Removed quota stream wrapper, allow soft quota
The quota stream wrapper was used to provide a hard quota that would
truncate files when exceeding the quota. Since we decided to go with a
soft quota, the stream wrapper isn't necessary any more.

This fixes issues where file would randomly be truncated when the quota
was almost reached, happening with encryption and could happen when
uploading multiple files from multiple sources concurrently.